pyproject.toml
[project]
name = "convo-summarizer"
version = "0.1.0"
description = "Scripts to transcribe and summarize conversations"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "assemblyai>=0.36.0",
    "openai-whisper>=20240930",
    "openai>=1.58.1",
    "pydub>=0.25.1",
    "python-dotenv>=1.0.1",
]
